+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add available var list in SDS Html Editor
 - Convert get supplier price function to return json instead of html
 - Improve privacy settings in ingredients - a message to let user know that if privacy is enabled, the ingredient data will not be available to other user in the local installation (This is a place holder for a future release)
+- Added PV Online in ingredient search provider
 
 ### Version 6.3
 - Add phpMyAdmin for docker images